Fixing after merging

+++ src/main/java/com/example/trading_system/service/Facade.java
@@ -93,31 +93,73 @@ public class Facade {
         }
     }
     public ResponseEntity<String> addProduct(String username, int product_id, String store_name, String product_name, String product_description,
-                                             double product_price, int product_quantity, double rating, Category category, List<String> keyWords){
-        return marketService.addProduct(username,product_id,store_name,product_name,product_description,product_price,product_quantity,rating,category,keyWords);
+                                             double product_price, int product_quantity, double rating, Category category, List<String> keyWords,String token){
+        if(Security.validateToken(token,username)) {
+            return marketService.addProduct(username, product_id, store_name, product_name, product_description, product_price, product_quantity, rating, category, keyWords);
+        }
+        else{
+            return new ResponseEntity<>("Expired token", HttpStatus.BAD_REQUEST);
+        }
     }
 
-    public ResponseEntity<String> removeProduct(String username, String store_name, int product_id){
-        return marketService.removeProduct(username,store_name,product_id);
+    public ResponseEntity<String> removeProduct(String username, String store_name, int product_id,String token){
+        if(Security.validateToken(token,username)) {
+            return marketService.removeProduct(username, store_name, product_id);
+        }
+        else{
+            return new ResponseEntity<>("Expired token", HttpStatus.BAD_REQUEST);
+        }
     }
 
-    public ResponseEntity<String> setProduct_name(String username, String store_name_id, int productId, String product_name){
-        return marketService.setProduct_name(username,store_name_id,productId,product_name);
+    public ResponseEntity<String> setProduct_name(String username, String store_name_id, int productId, String product_name,String token){
+        if(Security.validateToken(token,username)) {
+            return marketService.setProduct_name(username, store_name_id, productId, product_name);
+        }
+        else{
+            return new ResponseEntity<>("Expired token", HttpStatus.BAD_REQUEST);
+        }
     }
-    public ResponseEntity<String> setProduct_description(String username,String store_name_id,int productId,String product_description){
-        return marketService.setProduct_description(username,store_name_id,productId,product_description);
+    public ResponseEntity<String> setProduct_description(String username,String store_name_id,int productId,String product_description,String token){
+        if(Security.validateToken(token,username)) {
+            return marketService.setProduct_description(username, store_name_id, productId, product_description);
+        }
+        else{
+            return new ResponseEntity<>("Expired token", HttpStatus.BAD_REQUEST);
+        }
     }
-    public ResponseEntity<String> setProduct_price(String username,String store_name_id,int productId,int product_price){
-        return marketService.setProduct_price(username,store_name_id,productId,product_price);
+    public ResponseEntity<String> setProduct_price(String username,String store_name_id,int productId,int product_price,String token){
+        if(Security.validateToken(token,username)) {
+            return marketService.setProduct_price(username, store_name_id, productId, product_price);
+        }
+        else{
+            return new ResponseEntity<>("Expired token", HttpStatus.BAD_REQUEST);
+        }
     }
-    public ResponseEntity<String> setProduct_quantity(String username,String store_name_id,int productId,int product_quantity){
-        return marketService.setProduct_quantity(username,store_name_id,productId,product_quantity);
+    public ResponseEntity<String> setProduct_quantity(String username,String store_name_id,int productId,int product_quantity,String token){
+        if(Security.validateToken(token,username)) {
+            return marketService.setProduct_quantity(username, store_name_id, productId, product_quantity);
+        }
+        else{
+            return new ResponseEntity<>("Expired token", HttpStatus.BAD_REQUEST);
+        }
+
     }
-    public ResponseEntity<String> setRating(String username,String store_name_id,int productId,int rating){
-        return marketService.setRating(username,store_name_id,productId,rating);
+    public ResponseEntity<String> setRating(String username,String store_name_id,int productId,int rating,String token){
+        if(Security.validateToken(token,username)) {
+            return marketService.setRating(username, store_name_id, productId, rating);
+        }
+        else{
+            return new ResponseEntity<>("Expired token", HttpStatus.BAD_REQUEST);
+        }
+
     }
-    public ResponseEntity<String> setCategory(String username,String store_name_id,int productId,Category category){
-        return marketService.setCategory(username,store_name_id,productId,category);
+    public ResponseEntity<String> setCategory(String username,String store_name_id,int productId,Category category,String token){
+        if(Security.validateToken(token,username)) {
+            return marketService.setCategory(username, store_name_id, productId, category);
+        }
+        else{
+            return new ResponseEntity<>("Expired token", HttpStatus.BAD_REQUEST);
+        }
     }
 
     public String login(int id, String username, String password){
